# Resolve authorization errors and compiler warnings in Soroban contracts