--- a/contracts/course/course_access/src/functions/mod.rs
+++ b/contracts/course/course_access/src/functions/mod.rs
@@ -9,7 +9,6 @@ pub mod save_profile;
 pub mod transfer_course_access;
 
 pub use grant_access::*;
-pub use has_access::*;
 pub use list_course_access::course_access_list_course_access;
 pub use list_user_courses::course_access_list_user_courses;
 pub use revoke_access::*;